@@ -1,21 +1,27 @@
 module github.com/nerdzeu/nerdz-api
 
-go 1.14
+go 1.22
 
 require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/galeone/igor v1.0.0
-	github.com/google/uuid v1.1.1 // indirect
-	github.com/kr/text v0.2.0 // indirect
-	github.com/labstack/echo/v4 v4.1.16
-	github.com/niemeyer/pretty v0.0.0-20200227124842-a10e7caefd8e // indirect
+	github.com/galeone/igor v1.0.10
+	github.com/labstack/echo/v4 v4.11.4
 	github.com/openshift/osin v1.0.1
-	github.com/pborman/uuid v1.2.0 // indirect
-	github.com/rs/cors v1.7.0
-	github.com/stretchr/testify v1.5.1 // indirect
-	golang.org/x/crypto v0.0.0-20200323165209-0ec3e9974c59 // indirect
-	golang.org/x/net v0.0.0-20200324143707-d3edc9973b7e
-	golang.org/x/sys v0.0.0-20200331124033-c3d80250170d // indirect
-	gopkg.in/check.v1 v1.0.0-20200227125254-8fa46927fb4f // indirect
-	gopkg.in/yaml.v2 v2.2.8 // indirect
+	github.com/rs/cors v1.10.1
+	golang.org/x/net v0.22.0
+)
+
+require (
+	github.com/golang-jwt/jwt v3.2.2+incompatible // indirect
+	github.com/google/uuid v1.6.0 // indirect
+	github.com/labstack/gommon v0.4.2 // indirect
+	github.com/lib/pq v1.10.9 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/pborman/uuid v1.2.1 // indirect
+	github.com/valyala/bytebufferpool v1.0.0 // indirect
+	github.com/valyala/fasttemplate v1.2.2 // indirect
+	golang.org/x/crypto v0.21.0 // indirect
+	golang.org/x/sys v0.18.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+	golang.org/x/time v0.5.0 // indirect
 )
