@@ -23,6 +23,7 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/labstack/gommon/log"
 	"github.com/nerdzeu/nerdz-api/utils"
 )
 
@@ -142,26 +143,26 @@ func (post *ProjectPost) Language() string {
 
 // Revisions returns all the revisions of the message
 func (post *ProjectPost) Revisions() (modifications []string) {
-	Db().Model(ProjectPostRevision{}).Where(&ProjectPostRevision{Hpid: post.ID()}).Pluck("message", &modifications)
+	_ = Db().Model(ProjectPostRevision{}).Where(&ProjectPostRevision{Hpid: post.ID()}).Pluck("message", &modifications)
 	return
 }
 
 // RevisionsNumber returns the number of the revisions
 func (post *ProjectPost) RevisionsNumber() (count uint8) {
-	Db().Model(ProjectPostRevision{}).Where(&ProjectPostRevision{Hpid: post.ID()}).Count(&count)
+	_ = Db().Model(ProjectPostRevision{}).Where(&ProjectPostRevision{Hpid: post.ID()}).Count(&count)
 	return
 }
 
 // Votes returns the post's votes value
 func (post *ProjectPost) VotesCount() (sum int) {
-	Db().Model(ProjectPostVote{}).Select("COALESCE(sum(vote), 0)").Where(&ProjectPostVote{Hpid: post.ID()}).Scan(&sum)
+	_ = Db().Model(ProjectPostVote{}).Select("COALESCE(sum(vote), 0)").Where(&ProjectPostVote{Hpid: post.ID()}).Scan(&sum)
 	return
 }
 
 // Votes returns a pointer to a slice of Vote
 func (post *ProjectPost) Votes() *[]Vote {
 	ret := []ProjectPostVote{}
-	Db().Model(ProjectPostVote{}).Where(&ProjectPostVote{Hpid: post.ID()}).Scan(&ret)
+	_ = Db().Model(ProjectPostVote{}).Where(&ProjectPostVote{Hpid: post.ID()}).Scan(&ret)
 	var retVotes []Vote
 	for _, v := range ret {
 		vote := v
@@ -174,7 +175,7 @@ func (post *ProjectPost) Votes() *[]Vote {
 // Bookmarks returns a pointer to a slice of Bookmark
 func (post *ProjectPost) Bookmarks() *[]Bookmark {
 	ret := []ProjectPostBookmark{}
-	Db().Model(ProjectPostBookmark{}).Where(&ProjectPostBookmark{Hpid: post.ID()}).Scan(&ret)
+	_ = Db().Model(ProjectPostBookmark{}).Where(&ProjectPostBookmark{Hpid: post.ID()}).Scan(&ret)
 	var retBookmarks []Bookmark
 	for _, b := range ret {
 		bookmark := b
@@ -187,7 +188,7 @@ func (post *ProjectPost) Bookmarks() *[]Bookmark {
 // Lurks returns a pointer to a slice of Lurk
 func (post *ProjectPost) Lurks() *[]Lurk {
 	ret := []ProjectPostLurk{}
-	Db().Model(ProjectPostLurk{}).Where(&ProjectPostLurk{Hpid: post.ID()}).Scan(&ret)
+	_ = Db().Model(ProjectPostLurk{}).Where(&ProjectPostLurk{Hpid: post.ID()}).Scan(&ret)
 	var retLurkers []Lurk
 	for _, l := range ret {
 		lurker := l
@@ -200,7 +201,7 @@ func (post *ProjectPost) Lurks() *[]Lurk {
 // Locks returns a pointer to a slice of Lock
 func (post *ProjectPost) Locks() *[]Lock {
 	ret := []ProjectPostLock{}
-	Db().Model(ProjectPostLock{}).Where(&ProjectPostLock{Hpid: post.ID()}).Scan(&ret)
+	_ = Db().Model(ProjectPostLock{}).Where(&ProjectPostLock{Hpid: post.ID()}).Scan(&ret)
 	var retLockers []Lock
 	for _, l := range ret {
 		locker := l
@@ -217,7 +218,9 @@ func (post *ProjectPost) Comments(options CommentlistOptions) *[]ExistingComment
 
 	query := Db().Model(ProjectPostComment{})
 	query = commentlistQueryBuilder(query, options).Where(&ProjectPostComment{Hpid: post.ID()})
-	query.Scan(&comments)
+	if err := query.Scan(&comments); err != nil {
+		log.Errorf("(Postlist) Error in query.Scan: %s", err)
+	}
 
 	comments = utils.ReverseSlice(comments).([]ProjectPostComment)
 
@@ -231,7 +234,7 @@ func (post *ProjectPost) Comments(options CommentlistOptions) *[]ExistingComment
 
 // CommentsCount returns the number of comment's post
 func (post *ProjectPost) CommentsCount() (count uint8) {
-	Db().Model(ProjectPostComment{}).Where(&ProjectPostComment{Hpid: post.ID()}).Count(&count)
+	_ = Db().Model(ProjectPostComment{}).Where(&ProjectPostComment{Hpid: post.ID()}).Count(&count)
 	return
 }
 
@@ -247,7 +250,7 @@ func (*ProjectPost) Type() string {
 
 // NumericBookmarkers returns a slice of uint64 representing the ids of the users that bookmarked the post
 func (post *ProjectPost) NumericBookmarkers() (bookmarkers []uint64) {
-	Db().Model(ProjectPostBookmark{}).Where(&ProjectPostBookmark{Hpid: post.ID()}).Pluck(`"from"`, &bookmarkers)
+	_ = Db().Model(ProjectPostBookmark{}).Where(&ProjectPostBookmark{Hpid: post.ID()}).Pluck(`"from"`, &bookmarkers)
 	return
 }
 
@@ -258,13 +261,13 @@ func (post *ProjectPost) Bookmarkers() []*User {
 
 // BookmarksCount returns the number of users that bookmarked the post
 func (post *ProjectPost) BookmarksCount() (count uint8) {
-	Db().Model(ProjectPostBookmark{}).Where(&ProjectPostBookmark{Hpid: post.ID()}).Count(&count)
+	_ = Db().Model(ProjectPostBookmark{}).Where(&ProjectPostBookmark{Hpid: post.ID()}).Count(&count)
 	return
 }
 
 // NumericLurkers returns a slice of uint64 representing the ids of the users that lurked the post
 func (post *ProjectPost) NumericLurkers() (lurkers []uint64) {
-	Db().Model(ProjectPostLurk{}).Where(&ProjectPostLurk{Hpid: post.ID()}).Pluck(`"from"`, &lurkers)
+	_ = Db().Model(ProjectPostLurk{}).Where(&ProjectPostLurk{Hpid: post.ID()}).Pluck(`"from"`, &lurkers)
 	return
 }
 
@@ -275,7 +278,7 @@ func (post *ProjectPost) Lurkers() []*User {
 
 // LurkersCount returns the number of users that are lurking the post
 func (post *ProjectPost) LurkersCount() (count uint8) {
-	Db().Model(ProjectPostLurk{}).Where(&ProjectPostLurk{Hpid: post.ID()}).Count(&count)
+	_ = Db().Model(ProjectPostLurk{}).Where(&ProjectPostLurk{Hpid: post.ID()}).Count(&count)
 	return
 }
 
