@@ -23,6 +23,7 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/labstack/gommon/log"
 	"github.com/nerdzeu/nerdz-api/utils"
 )
 
@@ -124,14 +125,14 @@ func (post *UserPost) Owners() (ret []*User) {
 
 // Votes returns the post's votes value
 func (post *UserPost) VotesCount() (sum int) {
-	Db().Model(UserPostVote{}).Select("COALESCE(sum(vote), 0)").Where(&UserPostVote{Hpid: post.ID()}).Scan(&sum)
+	_ = Db().Model(UserPostVote{}).Select("COALESCE(sum(vote), 0)").Where(&UserPostVote{Hpid: post.ID()}).Scan(&sum)
 	return
 }
 
 // Votes returns a pointer to a slice of Vote
 func (post *UserPost) Votes() *[]Vote {
 	ret := []UserPostVote{}
-	Db().Model(UserPostVote{}).Where(&UserPostVote{Hpid: post.ID()}).Scan(&ret)
+	_ = Db().Model(UserPostVote{}).Where(&UserPostVote{Hpid: post.ID()}).Scan(&ret)
 	var retVotes []Vote
 	for _, v := range ret {
 		vote := v
@@ -144,7 +145,7 @@ func (post *UserPost) Votes() *[]Vote {
 // Bookmarks returns a pointer to a slice of Bookmark
 func (post *UserPost) Bookmarks() *[]Bookmark {
 	ret := []UserPostBookmark{}
-	Db().Model(UserPostBookmark{}).Where(&UserPostBookmark{Hpid: post.ID()}).Scan(&ret)
+	_ = Db().Model(UserPostBookmark{}).Where(&UserPostBookmark{Hpid: post.ID()}).Scan(&ret)
 	var retBookmarks []Bookmark
 	for _, b := range ret {
 		bookmark := b
@@ -157,7 +158,7 @@ func (post *UserPost) Bookmarks() *[]Bookmark {
 // Lurks returns a pointer to a slice of Lurk
 func (post *UserPost) Lurks() *[]Lurk {
 	ret := []UserPostLurk{}
-	Db().Model(UserPostLurk{}).Where(&UserPostLurk{Hpid: post.ID()}).Scan(&ret)
+	_ = Db().Model(UserPostLurk{}).Where(&UserPostLurk{Hpid: post.ID()}).Scan(&ret)
 	var retLurkers []Lurk
 	for _, l := range ret {
 		lurker := l
@@ -170,7 +171,7 @@ func (post *UserPost) Lurks() *[]Lurk {
 // Locks returns a pointer to a slice of Lock
 func (post *UserPost) Locks() *[]Lock {
 	ret := []UserPostLock{}
-	Db().Model(UserPostLock{}).Where(&UserPostLock{Hpid: post.ID()}).Scan(&ret)
+	_ = Db().Model(UserPostLock{}).Where(&UserPostLock{Hpid: post.ID()}).Scan(&ret)
 	var retLockers []Lock
 	for _, l := range ret {
 		locker := l
@@ -199,13 +200,13 @@ func (post *UserPost) Language() string {
 
 // Revisions returns all the revisions of the message
 func (post *UserPost) Revisions() (modifications []string) {
-	Db().Model(UserPostRevision{}).Where(&UserPostRevision{Hpid: post.ID()}).Pluck("message", &modifications)
+	_ = Db().Model(UserPostRevision{}).Where(&UserPostRevision{Hpid: post.ID()}).Pluck("message", &modifications)
 	return
 }
 
 // RevisionsNumber returns the number of the revisions
 func (post *UserPost) RevisionsNumber() (count uint8) {
-	Db().Model(UserPostRevision{}).Where(&UserPostRevision{Hpid: post.ID()}).Count(&count)
+	_ = Db().Model(UserPostRevision{}).Where(&UserPostRevision{Hpid: post.ID()}).Count(&count)
 	return
 }
 
@@ -216,7 +217,9 @@ func (post *UserPost) Comments(options CommentlistOptions) *[]ExistingComment {
 
 	query := Db().Model(UserPostComment{})
 	query = commentlistQueryBuilder(query, options).Where(&UserPostComment{Hpid: post.ID()})
-	query.Scan(&comments)
+	if err := query.Scan(&comments); err != nil {
+		log.Errorf("(UserPost::Comments) Error in query.Scan: %s", err)
+	}
 
 	comments = utils.ReverseSlice(comments).([]UserPostComment)
 
@@ -230,7 +233,7 @@ func (post *UserPost) Comments(options CommentlistOptions) *[]ExistingComment {
 
 // CommentsCount returns the number of comment's post
 func (post *UserPost) CommentsCount() (count uint8) {
-	Db().Model(UserPostComment{}).Where(&UserPostComment{Hpid: post.ID()}).Count(&count)
+	_ = Db().Model(UserPostComment{}).Where(&UserPostComment{Hpid: post.ID()}).Count(&count)
 	return
 }
 
@@ -246,7 +249,7 @@ func (*UserPost) Type() string {
 
 // NumericBookmarkers returns a slice of uint64 representing the ids of the users that bookmarked the post
 func (post *UserPost) NumericBookmarkers() (bookmarkers []uint64) {
-	Db().Model(UserPostBookmark{}).Where(&UserPostBookmark{Hpid: post.ID()}).Pluck(`"from"`, &bookmarkers)
+	_ = Db().Model(UserPostBookmark{}).Where(&UserPostBookmark{Hpid: post.ID()}).Pluck(`"from"`, &bookmarkers)
 	return
 }
 
@@ -257,13 +260,13 @@ func (post *UserPost) Bookmarkers() []*User {
 
 // BookmarksCount returns the number of users that bookmarked the post
 func (post *UserPost) BookmarksCount() (count uint8) {
-	Db().Model(UserPostBookmark{}).Where(&UserPostBookmark{Hpid: post.ID()}).Count(&count)
+	_ = Db().Model(UserPostBookmark{}).Where(&UserPostBookmark{Hpid: post.ID()}).Count(&count)
 	return
 }
 
 // NumericLurkers returns a slice of uint64 representing the ids of the users that lurked the post
 func (post *UserPost) NumericLurkers() (lurkers []uint64) {
-	Db().Model(UserPostLurk{}).Where(&UserPostLurk{Hpid: post.ID()}).Pluck(`"from"`, &lurkers)
+	_ = Db().Model(UserPostLurk{}).Where(&UserPostLurk{Hpid: post.ID()}).Pluck(`"from"`, &lurkers)
 	return
 }
 
@@ -274,7 +277,7 @@ func (post *UserPost) Lurkers() []*User {
 
 // LurkersCount returns the number of users that are lurking the post
 func (post *UserPost) LurkersCount() (count uint8) {
-	Db().Model(UserPostLurk{}).Where(&UserPostLurk{Hpid: post.ID()}).Count(&count)
+	_ = Db().Model(UserPostLurk{}).Where(&UserPostLurk{Hpid: post.ID()}).Count(&count)
 	return
 }
 
